requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.1
django-easy-pdf==0.1.1
django-filter==2.0.0
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.8
Markdown==3.0.1
Pillow==5.3.0
pycountry==18.5.26
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.5
reportlab==3.5.12
six==1.11.0
tinycss2==0.6.1
urllib3==1.24.1
WeasyPrint==43
webencodings==0.5.1
xhtml2pdf==0.2.3
